+++ src/jltypes.c
@@ -3439,6 +3439,24 @@ void post_boot_hooks(void)
     }
     export_small_typeof();
 }
+
+void post_image_load_hooks(void) {
+    // Ensure that `Base` has been loaded.
+    assert(jl_base_module != NULL);
+
+    jl_libdl_module = (jl_module_t *)jl_get_global(
+        ((jl_module_t *)jl_get_global(jl_base_module, jl_symbol("Libc"))),
+        jl_symbol("Libdl")
+    );
+    jl_libdl_dlopen_func = jl_get_global(
+        jl_libdl_module,
+        jl_symbol("dlopen")
+    );
+    jl_lazy_library_type = (jl_datatype_t *)jl_get_global(
+        jl_libdl_module,
+        jl_symbol("LazyLibrary")
+    );
+}
 #undef XX
 
 #ifdef __cplusplus